{{cookiecutter.repo_name}}/requirements.txt
# Core
pytz==2018.4
Django==2.0.6
django-configurations==2.0
gunicorn==19.8.1
newrelic==3.2.2.94

# For the persistence stores
psycopg2-binary==2.7.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==1.1.0

djangorestframework-jwt
cryptography

# Developer Tools
ipdb==0.11
ipython==6.4.0
mkdocs==0.17.4
flake8==3.5.0
mypy==0.630

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1

